plant-stress/requirements.txt
# Core ML libraries
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
tensorboard>=2.13.0

# Geospatial (optional - for later phases)
geopandas>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.5.0
rtree>=1.0.0
rasterio>=1.3.0

# ONNX and optimization
onnx>=1.14.0
onnxruntime>=1.15.0
coremltools>=7.0.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
requests>=2.31.0

# (Removed GPS & Web deps to keep runtime minimal)

# TensorFlow (optional baseline)
tensorflow>=2.14.0,<3.0.0

# Kaggle CLI (for dataset downloads)
kaggle>=1.5.0
